requirements.txt

Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
requests==2.32.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
django-jazzmin==2.6.0

#Django==5.0.6

# channels==4.1.0
# channels-redis==4.2.0
# charset-normalizer==3.3.2
# constantly==23.10.4
# cryptography==42.0.8
# daphne==4.1.2
# Django==4.2.13
# django-cors-headers==4.3.1
# djangorestframework==3.15.1
# hiredis==2.3.2
# hyperlink==21.0.0
# idna==3.7
# incremental==22.10.0
# msgpack==1.0.8
# pkg_resources==0.0.0
# pyasn1==0.6.0
# pyasn1_modules==0.4.0
# pycparser==2.22
# pyOpenSSL==24.1.0
# redis==5.0.6
# redis-server==6.0.9
# requests==2.32.3
# service-identity==24.1.0
# six==1.16.0
# sqlparse==0.5.0
# Twisted==24.3.0
# txaio==23.1.1
# typing_extensions==4.12.2
# urllib3==2.2.2
# zope.interface==6.4.post2
